RDKB-44328 : Exclude telemetry reports with SNMP/TR181/GUI reboots
Reason for change: Reboot from GUI should complete within 30s .
 Log upload or telemetry operation adds significant delays and should not be included
 with reboots from GUI or SNMP .
Test Procedure: 1] run below command to trigger reboot via TR181
 dmcli eRT setv Device.X_CISCO_COM_DeviceControl.RebootDevice string Device
2] Ensure device reboots within 30s after above command is initiated.
3] Check the logs in telemetry telemetry2_0.txt.0 and make sure belwo logging is available -
   telemetry reports excluded to avoid delays in reboot from GUI/TR181/SNMP triggers
Risks: Medium
Priority: P0

(cherry picked from commit 32b5fa7c48cb018dd30efd8d34e50578325c7f54)

+++ backupLogs.sh
@@ -246,19 +246,24 @@ backupLogsonReboot_nvram2()
         fi 
 
         echo_t  "[DEBUG] $0 Notify telemetry to execute now before log upload !!!" >> /rdklogs/logs/telemetry2_0.txt.0
-        sh /lib/rdk/dca_utility.sh 2 &
-        local loop=0
-        while :
-        do
-            sleep 10
-            loop=$((loop+1))
-            if [ -f "$DCA_COMPLETED" ] || [ "$loop" -ge 6 ]
-            then
-                echo_t "[DEBUG] $0 telemetry operation completed loop count = $loop" >> /rdklogs/logs/telemetry2_0.txt.0
-                rm -rf $DCA_COMPLETED
-                break
-            fi
-        done
+	if [ "$needReboot" = "true" ]; then
+            # Similar to RDKB-9204 - Avoid telemetry computation from GUI or SNMP initiated reboot
+            echo_t  "[DEBUG] $0 telemetry reports excluded to avoid delays in reboot from SNMP/TR181/GUI !!!" >> /rdklogs/logs/telemetry2_0.txt.0
+	else 
+            sh /lib/rdk/dca_utility.sh 2 &
+            local loop=0
+            while :
+            do
+                sleep 10
+                loop=$((loop+1))
+                if [ -f "$DCA_COMPLETED" ] || [ "$loop" -ge 6 ]
+                then
+                    echo_t "[DEBUG] $0 telemetry operation completed loop count = $loop" >> /rdklogs/logs/telemetry2_0.txt.0
+                    rm -rf $DCA_COMPLETED
+                    break
+                fi
+            done
+        fi
 
 	syncLogs_nvram2
